Using Options to check if Dynamic worker resolution is enabled
