+++ test/WebJobs.Script.Tests/Extensions/EnvironmentExtensionsTests.cs
@@ -553,59 +553,5 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Extensions
 
             Assert.Equal(expected, env.IsInValidationMode());
         }
-
-        [Theory]
-        [InlineData(null, "node", true)]
-        [InlineData(null, "java|node", true)]
-        [InlineData(null, "", false)]
-        [InlineData(null, "| ", false)]
-        [InlineData(null, null, false)]
-        [InlineData(ScriptConstants.FeatureFlagDisableDynamicWorkerResolution, "node", false)]
-        [InlineData(ScriptConstants.FeatureFlagDisableDynamicWorkerResolution, "java|node", false)]
-        [InlineData(ScriptConstants.FeatureFlagDisableDynamicWorkerResolution, "| ", false)]
-
-        public void IsDynamicWorkerResolutionEnabled_HostingConfigAndFeatureFlags_WorksAsExpected(string featureFlagValue, string hostingConfigSetting, bool expected)
-        {
-            var mockConfiguration = new Mock<IConfiguration>();
-            var mockScriptHostManager = new Mock<IScriptHostManager>();
-
-            var hostingOptions = new FunctionsHostingConfigOptions();
-            hostingOptions.Features.Add(RpcWorkerConstants.WorkersAvailableForDynamicResolution, hostingConfigSetting);
-
-            var testEnvironment = new TestEnvironment();
-            testEnvironment.SetEnvironmentVariable(AzureWebJobsFeatureFlags, featureFlagValue);
-
-            var optionsMonitor = WorkerConfigurationResolverTestsHelper.GetTestWorkerConfigurationResolverOptions(mockConfiguration.Object, testEnvironment, mockScriptHostManager.Object, new OptionsWrapper<FunctionsHostingConfigOptions>(hostingOptions));
-
-            bool result = testEnvironment.IsDynamicWorkerResolutionEnabled(optionsMonitor);
-
-            Assert.Equal(expected, result);
-        }
-
-        [Theory]
-        [InlineData("node", "node", null, true)]
-        [InlineData("node", "java", null, false)]
-        [InlineData("java|node", null, null, true)]
-        [InlineData("node", "node", "workflowapp", true)]
-        [InlineData("java|node", null, "workflowapp", true)]
-        [InlineData("| ", null, "workflowapp", false)]
-        public void IsDynamicWorkerResolutionEnabled_WorkerRuntimeAndMultiLanguage_WorksAsExpected(string hostingConfigSetting, string workerRuntime, string multilanguageApp, bool expected)
-        {
-            var mockConfiguration = new Mock<IConfiguration>();
-            var mockScriptHostManager = new Mock<IScriptHostManager>();
-
-            var hostingOptions = new FunctionsHostingConfigOptions();
-            hostingOptions.Features.Add(RpcWorkerConstants.WorkersAvailableForDynamicResolution, hostingConfigSetting);
-
-            var testEnvironment = new TestEnvironment();
-            testEnvironment.SetEnvironmentVariable(AppKind, multilanguageApp);
-            testEnvironment.SetEnvironmentVariable(FunctionWorkerRuntime, workerRuntime);
-
-            var optionsMonitor = WorkerConfigurationResolverTestsHelper.GetTestWorkerConfigurationResolverOptions(mockConfiguration.Object, testEnvironment, mockScriptHostManager.Object, new OptionsWrapper<FunctionsHostingConfigOptions>(hostingOptions));
-
-            bool result = testEnvironment.IsDynamicWorkerResolutionEnabled(optionsMonitor);
-
-            Assert.Equal(expected, result);
-        }
     }
 }
